+++ composeApp/src/commonMain/kotlin/com/lastaosi/mycat/domain/usecase/CalculateAgeMonthUseCase.kt
@@ -8,11 +8,16 @@ import kotlin.time.ExperimentalTime
 class CalculateAgeMonthUseCase {
     @OptIn(ExperimentalTime::class)
     operator fun invoke(birthDate: String): Int {
-        val parts = birthDate.split("-")
-        val birthYear = parts[0].toInt()
-        val birthMonth = parts[1].toInt()
-        val now = Clock.System.now()
-            .toLocalDateTime(TimeZone.currentSystemDefault())
-        return (now.year - birthYear) * 12 + (now.monthNumber - birthMonth)
+        return try{
+            val parts = birthDate.split("-")
+            val birthYear = parts[0].toInt()
+            val birthMonth = parts[1].toInt()
+            val now = Clock.System.now()
+                .toLocalDateTime(TimeZone.currentSystemDefault())
+            (now.year - birthYear) * 12 + (now.monthNumber - birthMonth)
+        }catch (e: Exception){
+            0
+        }
+
     }
 }